+++ reduction.h
@@ -17,14 +17,14 @@ int generate_photometry_dft_data(datafile *data,
                                  double *ratio_mean, double *ratio_std, double *mma_mean, double *mma_std,
                                  double **freq, double **ampl, size_t *num_dft);
 
-int create_flat(const char *pattern, int minmax, const char *masterdark, const char *outname);
-int create_dark(const char *pattern, int minmax, const char *outname);
+int create_flat(const char *pattern, size_t minmax, const char *masterdark, const char *outname);
+int create_dark(const char *pattern, size_t minmax, const char *outname);
 
 int reduce_single_frame(char *framePath, char *darkPath, char *flatPath, char *outPath);
 int update_reduction(char *dataPath);
 int create_reduction_file(char *filePath);
 int update_preview(char *preview_filename, char *ds9_title);
 int calculate_bjd(char *date, char *time, char *ra_string, char *dec_string, double epoch);
-int create_ts(char *reference_date, char *reference_time, char **filenames, int num_datafiles, char *ts_filename);
+int create_ts(char *reference_date, char *reference_time, char **filenames, size_t num_datafiles, char *ts_filename);
 
 #endif
